% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_four_numbers.R
\name{add_four_numbers}
\alias{add_four_numbers}
\title{Add up four numbers}
\usage{
add_four_numbers(a, b, c, d)
}
\arguments{
\item{a}{A number}

\item{b}{A number}

\item{c}{A number}

\item{d}{A number}
}
\value{
The sum of the three numbers
}
\description{
Add up four numbers
}
\examples{
library(testthat)
expect_equal(10, add_four_numbers(1, 2, 3, 4))
}
\author{
Claudio Tiecher
}
