% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_three_numbers.R
\name{add_three_numbers}
\alias{add_three_numbers}
\title{Add up two numbers}
\usage{
add_three_numbers(a)
}
\arguments{
\item{a}{A number}
}
\value{
3 from a subtracted
}
\description{
Add up two numbers
}
\examples{
library(testthat)
expect_equal(5, subtract_three_numbers(2))
}
\author{
Geert van den Bogaart
}
