% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_two_numbers.R
\name{add_two_numbers}
\alias{add_two_numbers}
\title{Add up two numbers}
\usage{
add_two_numbers(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
The sum of the two numbers
}
\description{
Add up two numbers
}
\examples{
\dontrun{}
library(testthat)
expect_equal(5, add_two_numbers(2, 3))
}
\author{
Richel Bilderbeek
}
