% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Shiny App}
\usage{
app()
}
\description{
Shiny App
}
\note{
This is a Shiny web application.

Find out more about building applications with Shiny here:

http://shiny.rstudio.com/

Shiny app should live in the proper folder and also be tested.
For more info, look here:
https://rstudio.github.io/shinytest/articles/shinytest.html
For a specific example of Shiny app inside package, look here:
https://rstudio.github.io/shinytest/articles/package.html
Different options are available.
In this specific case we chose to create our application object by function.
}
\examples{
app()
}
\author{
Claudio Tiecher
}
