% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_png_filename.R
\name{check_png_filename}
\alias{check_png_filename}
\title{Check png filename}
\usage{
check_png_filename(png_filename)
}
\arguments{
\item{png_filename}{name of the PNG that the map will be saved to}
}
\description{
Check png filename
}
\author{
Alexine de Wit
}
