% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_show_bridge_openings.R
\name{check_show_bridge_openings}
\alias{check_show_bridge_openings}
\title{Check the value of `show_bridge_openings`}
\usage{
check_show_bridge_openings(a)
}
\arguments{
\item{show_bridge_openings}{character string indicating whether or not
bridge openings are plotted on top of the basemap.}
}
\description{
Check the value of `show_bridge_openings`
}
