% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{Create map}
\usage{
create_map(png_filename, show_bridge_openings = TRUE)
}
\arguments{
\item{png_filename}{name of the PNG that the map will be saved to}
}
\description{
Fetch a map of Groningen from OpenStreetMap
and save it as a PNG-file.
}
