% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map_traffic_densities.R
\name{create_map_traffic_densities}
\alias{create_map_traffic_densities}
\title{Create map with car traffic densities.}
\usage{
create_map_traffic_densities(png_filename, show_measurements = TRUE)
}
\arguments{
\item{png_filename}{name of the PNG that the map will be saved to}

\item{show_measurements}{Logical indicating whether or not to plot
car traffic densities on top of the base map}
}
\description{
Fetch a map of Groningen from OpenStreetMap,
add car traffic densities and save it as a PNG-file.
}
