% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map_viewer.R
\name{create_map_viewer}
\alias{create_map_viewer}
\title{Create map viewer}
\usage{
create_map_viewer(show_bridge_openings = TRUE)
}
\arguments{
\item{show_bridge_openings}{boolean indicating whether or not to
show bridge openings on top of base map}
}
\description{
Create an interactive map viewer.
}
