% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map_measurements.R
\name{create_map_with_car_traffic_densities}
\alias{create_map_with_car_traffic_densities}
\title{Create map}
\usage{
create_map_with_car_traffic_densities(png_filename, show_measurements = TRUE)
}
\arguments{
\item{png_filename}{name of the PNG that the map will be saved to}
}
\description{
Fetch a map of Groningen from OpenStreetMap
and save it as a PNG-file.
}
