% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_three_numbers.R
\name{divide_three_numbers}
\alias{divide_three_numbers}
\title{Divide 3 numbers}
\usage{
divide_three_numbers(a, b, c)
}
\arguments{
\item{a}{A number}

\item{b}{A number}

\item{c}{A number}
}
\value{
The division of 3 numbers
}
\description{
Divide 3 numbers
}
\examples{
library(testthat)
expect_equal(1, divide_three_numbers((10 / 5) / 2))
}
\author{
lizzie
}
