% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alexines_github_username.R
\name{get_alexines_github_username}
\alias{get_alexines_github_username}
\title{Get Alexines GitHub username}
\usage{
get_alexines_github_username()
}
\value{
Alexines GitHub username
}
\description{
Get Alexines GitHub username
}
\examples{
get_alexines_github_username()
}
\author{
Alexine
}
