% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annes_github_username.R
\name{get_annes_github_username}
\alias{get_annes_github_username}
\title{Anne's GitHub username}
\usage{
get_annes_github_username()
}
\value{
Anne's GitHub username
}
\description{
Anne's GitHub username
}
\examples{
\dontrun{}
library(testthat)

username <- get_annes_github_username()
expect_equal("fastnotfurious", username)
}
\author{
Anne-Kathrin Kleine
}
