% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bridge_openings.R
\name{get_bridge_openings}
\alias{get_bridge_openings}
\title{Get all the real time bridge openings in the Netherlands}
\usage{
get_bridge_openings()
}
\description{
Get all the real time bridge openings in the Netherlands
}
