% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_measurements_fast.R
\name{get_car_densities}
\alias{get_car_densities}
\title{Get current measurements of traffic}
\usage{
get_car_densities()
}
\description{
Get current measurements of traffic
}
\examples{
# Put example code here
download the positions of the measurement loops from NDW site
unzip the file
value_query <-".//d1:measurementSiteName/d1:values/d1:value"
value<-xml2::xml_find_first(parents, value_query) \%>\% xml2::xml_text()
now download the second file with all the traffic densities
unzip this file as well
now extract the location names and the traffic densities
the following part is still incorrect,
because it finds one more location than density
and because it only finds the density on one position (rijbaan, weghelft)
this part is terribly slow
if the whole thing is done (with for(i in site_names_1)),
that is why I currently only check the first 1000

}
