% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_claudios_github_username.R
\name{get_claudios_github_username}
\alias{get_claudios_github_username}
\title{Get Claudio's GitHub username}
\usage{
get_claudios_github_username()
}
\value{
Claudio's GitHub username
}
\description{
Get Claudio's GitHub username
}
\examples{
get_claudios_github_username()
}
\author{
Claudio Tiecher
}
