% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coordinates.R
\name{get_coordinates}
\alias{get_coordinates}
\title{Collect GPS coordinates from the NDW website}
\usage{
get_coordinates(category = "measurements")
}
\arguments{
\item{category}{specify category (measurements,...)}
}
\description{
This function fetches GPS coordinates from online sources:
measurements: http://opendata.ndw.nu/measurement_current.xml.gz
Add others here ...
}
\examples{
\dontrun{get_coordinates()}
\dontrun{get_coordinates("measurements")}

}
\author{
Claudio Tiecher
}
