% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frans_github_username.R
\name{get_frans_github_username}
\alias{get_frans_github_username}
\title{Get Frans's GitHub username}
\usage{
get_frans_github_username()
}
\value{
Frans's GitHub username
}
\description{
Get Frans's GitHub username
}
\examples{
get_frans_github_username()
}
\author{
Frans
}
