% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geerts_github_username.R
\name{get_geerts_github_username}
\alias{get_geerts_github_username}
\title{Get Geert's GitHub username}
\usage{
get_geerts_github_username()
}
\value{
Geert's GitHub username
}
\description{
Get Geert's GitHub username
}
\examples{
\dontrun{}
library(testthat)

username <- get_geerts_github_username()
expect_equal("bogeert", username)
}
\author{
Geert van den Bogaart
}
