% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jans_github_username.R
\name{get_jans_github_username}
\alias{get_jans_github_username}
\title{Get Jan's GitHub username}
\usage{
get_jans_github_username()
}
\value{
Jan's GitHub username
}
\description{
Get Jan's GitHub username
}
\examples{
get_jans_github_username()
}
\author{
Jan Derk Kotlarski
}
