% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maxims_github_username.R
\name{get_maxims_github_username}
\alias{get_maxims_github_username}
\title{Get Maxim's GitHub username}
\usage{
get_maxims_github_username()
}
\value{
Maxim's GitHub username
}
\description{
Get Maxim's GitHub username
}
\examples{
library(testthat)

username <- get_maxims_github_username()
expect_equal("MaximBaranov87", username)
}
\author{
Richel Bilderbeek
}
