% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pedros_github_username.R
\name{get_pedros_github_username}
\alias{get_pedros_github_username}
\title{Get Pedro's GitHub username}
\usage{
get_pedros_github_username()
}
\value{
Pedro's GitHub username
}
\description{
Get Pedro's GitHub username
}
\examples{
\dontrun{}
username <- get_pedros_github_username()
testthat::expect_equal("Neves-P", username)
}
\author{
Pedro Neves
}
