% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stefanos_github_username.R
\name{get_stefanos_github_username}
\alias{get_stefanos_github_username}
\title{Get Stefano's GitHub username}
\usage{
get_stefanos_github_username()
}
\value{
Stefano's GitHub username
}
\description{
Get Stefano's GitHub username
}
\examples{
library(testthat)

username <- get_stefanos_github_username()
expect_equal("swom", username)
}
\author{
Stefano Tiso
}
