% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_willys_github_username.R
\name{get_willys_github_username}
\alias{get_willys_github_username}
\title{Get Willy's GitHub username}
\usage{
get_willys_github_username()
}
\value{
Willy's GitHub username
}
\description{
Get Willy's GitHub username
}
\examples{
get_willys_github_username()
}
\author{
Willy Tadema
}
