% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xml.R
\name{get_xml}
\alias{get_xml}
\title{Download zipped XML-file and convert it to an XML document}
\usage{
get_xml(url)
}
\arguments{
\item{url}{Download URL of the zipped XML-file}
}
\value{
XML document
}
\description{
Download zipped XML-file and convert it to an XML document
}
\examples{
get_xml("http://opendata.ndw.nu/brugopeningen.xml.gz")
}
