% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xml_from_url.R
\name{get_xml_from_url}
\alias{get_xml_from_url}
\title{Download the xml file found at the given url}
\usage{
get_xml_from_url(url)
}
\arguments{
\item{url}{where to find the xml file}
}
\description{
Download the xml file found at the given url
}
\examples{
\dontrun{get_xml_from_url("http://opendata.ndw.nu/brugopeningen.xml.gz")}
}
