% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informalr-package.R
\docType{package}
\name{informalr-package}
\alias{informalr}
\alias{informalr-package}
\title{informalr: Learn R Informally in a Team}
\description{
There are multiple ways to learn R.
    One way is to join the informalr team.
    There, one can learn R informally in a team.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/richelbilderbeek/informalr/}
  \item Report bugs at \url{https://github.com/richelbilderbeek/informalr/issues}
}

}
\author{
\strong{Maintainer}: Richèl J.C. Bilderbeek \email{richel@richelbilderbeek.nl} (\href{https://orcid.org/0000-0003-1107-7049}{ORCID})

}
\keyword{internal}
