% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulo_two_numbers.R
\name{modulo_two_numbers}
\alias{modulo_two_numbers}
\title{Get the modulo of two numbers}
\usage{
modulo_two_numbers(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
The modulo of \code{a} and \code{b}
}
\description{
Get the modulo of two numbers
}
\examples{
\dontrun{}
modulo_two_numbers(4, 2)
modulo_two_numbers(10, 5)
}
\author{
Willy Tadema
}
