% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply_four_numbers.R
\name{multiply_four_numbers}
\alias{multiply_four_numbers}
\title{Multiply four numbers}
\usage{
multiply_four_numbers(a, b, c, d)
}
\arguments{
\item{a}{A number}

\item{b}{A number}

\item{c}{A number}

\item{d}{A number}
}
\value{
The product of the four numbers
}
\description{
Multiply four numbers
}
\examples{
multiply_four_numbers(1, 2, 3, 4)
}
\author{
Alexine
}
