% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bridge_openings.R
\name{plot_bridge_openings}
\alias{plot_bridge_openings}
\title{Create map}
\usage{
plot_bridge_openings(png_filename)
}
\arguments{
\item{png_filename}{name of the PNG that the map will be saved to}
}
\description{
Fetch a map of the province (state) Groningen from OpenStreetMap
Fetch bridge openings using function get_bridge_openings
and save it as a PNG-file.
}
