% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_two_numbers.R
\name{power_two_numbers}
\alias{power_two_numbers}
\title{Give the power of one number to the second one}
\usage{
power_two_numbers(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
Number A to the power of number B
}
\description{
Give the power of one number to the second one
}
\examples{
power_two_numbers(3, 4)
}
\author{
Jan Derk Kotlarski
}
