% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_four_numbers.R
\name{subtract_four_numbers}
\alias{subtract_four_numbers}
\title{Subtract four numbers}
\usage{
subtract_four_numbers(a)
}
\arguments{
\item{a}{A number}
}
\value{
4 from a subtracted
}
\description{
Subtract four numbers
}
\examples{
subtract_four_numbers(5)
}
\author{
Geert van den Bogaart
}
