% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_three_numbers.R
\name{subtract_three_numbers}
\alias{subtract_three_numbers}
\title{Subtract three numbers}
\usage{
subtract_three_numbers(a)
}
\arguments{
\item{a}{A number}
}
\value{
3 from a subtracted
}
\description{
Subtract three numbers
}
\examples{
\dontrun{}
library(testthat)
expect_equal(2, subtract_three_numbers(5))
}
\author{
Geert van den Bogaart
}
