% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_three_numbers.R
\name{add_three_numbers}
\alias{add_three_numbers}
\title{Add up three numbers}
\usage{
add_three_numbers(a, b, c)
}
\arguments{
\item{a}{A number}

\item{b}{A number}

\item{c}{A number}
}
\value{
The sum of the three numbers
}
\description{
Add up three numbers
}
\examples{
\dontrun{}
library(testthat)
expect_equal(6, add_three_numbers(1, 2, 3))
}
\author{
Anne-Kathrin Kleine
}
