% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{Create map}
\usage{
create_map(
  png_filename,
  show_bridge_openings = TRUE,
  right_lon = 1,
  left_lon = 2,
  top_lat = 3,
  bottom_lat = 4
)
}
\arguments{
\item{png_filename}{name of the PNG that the map will be saved to}

\item{show_bridge_openings}{boolean indicating whether or not to
show bridge openings on top of base map}

\item{right_lon}{right longitude}

\item{left_lon}{left longitude}

\item{top_lat}{top latitude}

\item{bottom_lat}{bottom latitude}
}
\description{
Fetch a map of Groningen from OpenStreetMap
and save it as a PNG-file.
}
