% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{Create map}
\usage{
create_map(png_filename, show_bridge_openings = TRUE, fake_data = FALSE)
}
\arguments{
\item{png_filename}{name of the PNG that the map will be saved to}

\item{show_bridge_openings}{Add layer with bridge openings
on top of the base map}

\item{fake_data}{Indicates whether to use test data}
}
\description{
Fetch a map of Groningen from OpenStreetMap
and save it as a PNG-file.
}
