% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{Create map}
\usage{
create_map(
  png_filename,
<<<<<<< HEAD
  show_bridge_openings = TRUE,
  fake_data = FALSE,
  show_car_traffic_density = FALSE,
=======
  show_bridge_openings = "yes",
>>>>>>> develop
  right_lon = 1,
  left_lon = 2,
  top_lat = 3,
  bottom_lat = 4
)
}
\arguments{
\item{png_filename}{name of the PNG that the map will be saved to}

\item{show_bridge_openings}{character string indicating whether or not
bridge openings are plotted on top of the basemap. The default,
\code{"yes"} adds the real time bridge openings to the map. \code{"fake"}
adds one fake bridge opening in the center of Groningen. \code{"no"}
does not add any bridge openings to the map.}

<<<<<<< HEAD
\item{fake_data}{Indicates whether to use test data}

\item{show_car_traffic_density}{Add layer of car traffic density}

\item{right_lon}{right longetitude}
=======
\item{right_lon}{right longitude}
>>>>>>> develop

\item{left_lon}{left longitude}

\item{top_lat}{top latitude}

<<<<<<< HEAD
\item{bottom_lat}{bottom lattitude
on top of the base map}
=======
\item{bottom_lat}{bottom latitude}
>>>>>>> develop
}
\description{
Fetch a map of Groningen from OpenStreetMap
and save it as a PNG-file.
}
