% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{Create a map}
\usage{
create_map(png_filename)
}
\arguments{
\item{png_filename}{name of the PNG that the map will be saved to}
}
\description{
Creates a map with the desired hiking info.
The map is saved as a PNG file.
}
