% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map_viewer.R
\name{create_map_viewer}
\alias{create_map_viewer}
\title{Create map viewer}
\usage{
create_map_viewer(show_bridge_openings = "yes")
}
\arguments{
\item{show_bridge_openings}{character string indicating whether or not
bridge openings are plotted on top of the basemap. The default,
\code{"yes"} adds the real time bridge openings to the map. \code{"fake"}
adds one fake bridge opening in the center of Groningen. \code{"no"}
does not add any bridge openings to the map.}
}
\description{
Create an interactive map viewer.
}
