% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_two_numbers.R
\name{divide_two_numbers}
\alias{divide_two_numbers}
\title{Divide two numbers}
\usage{
divide_two_numbers(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
The division of number `a` by number `b`.
}
\description{
Divide two numbers
}
\examples{
testthat::expect_equal(
  divide_two_numbers(a = 2, b = 5),
  0.4
)
}
