% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bridge_openings.R
\name{get_bridge_openings}
\alias{get_bridge_openings}
\title{Get real time bridge openings}
\usage{
get_bridge_openings(fake_data = FALSE)
}
\arguments{
\item{fake_data}{Indicates whether to return test data.
(GPS location of the Martini tower.)}
}
\description{
Get real time bridge openings
}
\examples{
df <- get_bridge_openings()
}
