% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bridge_openings.R
\name{get_bridge_openings}
\alias{get_bridge_openings}
\title{Get real time bridge openings}
\usage{
get_bridge_openings()
}
\description{
Get real time bridge openings
}
\examples{
get_bridge_openings()
}
