% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_car_densities.R,
%   R/get_current_measurements_fast.R
\name{get_car_densities}
\alias{get_car_densities}
\title{Get current measurements of traffic}
\usage{
get_car_densities()

get_car_densities()
}
\value{

}
\description{
Get current measurements of traffic

Get current measurements of traffic
}
\examples{
\dontrun{get_car_densities()}
# Put example code here
}
