% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_measurements_fast.R
\name{get_car_densities}
\alias{get_car_densities}
\title{Get current measurements of traffic}
\usage{
get_car_densities()
}
\description{
Get current measurements of traffic
}
\examples{
# Put example code here
}
