% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_richels_github_username.R
\name{get_richels_github_username}
\alias{get_richels_github_username}
\title{Get Richel's GitHub username}
\usage{
get_richels_github_username()
}
\value{
Richel's GitHub username
}
\description{
Get Richel's GitHub username
}
\examples{
\dontrun{}
library(testthat)

username <- get_richels_github_username()
expect_equal("richelbilderbeek", username)
}
\author{
Richel
}
