% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ic50s.R
\name{read_ic50s}
\alias{read_ic50s}
\title{Read demo data}
\usage{
read_ic50s(filename = "2016-third-quarter-stolen-guns.csv")
}
\arguments{
\item{filename}{The name of the file in
\code{inst/extdata} with the demo data}
}
\value{
A dataframe with the demo data.
}
\description{
Read the file in \code{inst/extdata}
with the demo data.
}
\examples{
data <- read_ic50s()
}
