% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ic50s.R
\name{read_ic50s}
\alias{read_ic50s}
\title{Read demo data}
\usage{
read_ic50s()
}
\value{
A dataframe with the contents of the \code{ic50s.csv} file.
}
\description{
Read the \code{ic50s.csv} file in \code{inst/extdata}.
}
\examples{
data <- read_ic50s()
}
